images/Load_Data_Python/requirements.txt
mysqlclient ~= 2.1.0
mysql-connector-python ~= 8.0.5
pandas